Валидаторы

+++ src/main/java/ru/rustam/catalog/dto/CatalogDto.java
@@ -3,13 +3,15 @@ package ru.rustam.catalog.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class CatalogDto {
-    private int id;
+    private Integer id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
 }
